add casting in compareTo method. It's campatible to JAVA 1.4 and fix templete complain for JAVA 1.5
git-svn-id: https://svn.potix.com/svn/zk1/trunk@3831 dd50bd9b-9913-0410-b9ba-e07a3075be86

+++ zcommon/src/org/zkoss/lang/Objects.java
@@ -230,7 +230,7 @@ public class Objects {
 		if (a == b || (a != null && b != null && a.equals(b)))
 			return true;
 		if ((a instanceof BigDecimal) && (b instanceof BigDecimal))
-			return ((BigDecimal)a).compareTo(b) == 0;
+			return ((BigDecimal)a).compareTo((BigDecimal) b) == 0;
 
 		if (a == null || !a.getClass().isArray())
 			return false;
